+++ R/mod_hierarchical_count_table.R
@@ -6,13 +6,22 @@ EC <- poc( # nolint
     MIN_PERCENT = "min_percent",
     TAB_DOWNLOAD = "table_download"
   ),
+  LBL = poc(
+    HIERARCHY = "Event count by",
+    GRP = "Group by",
+    MIN_PERCENT = "Minimum %"
+  ),
+  INFO = poc(
+    HIERARCHY = "Up to 2 selections allowed"
+  ),
   MSG = poc(
     VALIDATE = poc(
       NO_GRP = "No group selected",
-      NO_HIERARCHY = "No hierarchy selected or more than two levels selected",
+      NO_HIERARCHY = "No hierarchy selected",
       NO_MIN_PERCENT = "No minimum percent selected",
       NO_TABLE_ROWS = "Table dataset has 0 rows",
-      NO_POP_ROWS = "Population dataset has 0 rows"
+      NO_POP_ROWS = "Population dataset has 0 rows",
+      GRP_CLASH = "Group selection cannot be used in hierarchy"
     )
   )
 )
@@ -26,7 +35,7 @@ EC <- poc( # nolint
 #' @param pop_df `data.frame`
 #' A data frame containing the population data. It must have columns corresponding to subjects and group variables.
 #'
-#' @param hierarchy `character(1)`
+#' @param hierarchy `character(1|2)`
 #' A character vector of column names from `event_df` to use as the hierarchy. Can be one or two levels.
 #'
 #' @param group_var `character(1)`
@@ -64,23 +73,27 @@ compute_events_table <- function(event_df, pop_df, hierarchy, group_var, subjid_
   checkmate::assert_subset(group_var, names(pop_df))
   checkmate::assert_string(subjid_var, min.chars = 1)
 
+  # Replace NA values in group var factor with "<NA>" and add associated level
+  no_na_pop_df <- pop_df
+  no_na_pop_df[[group_var]] <- add_na_factor_level(no_na_pop_df[[group_var]])
+
   n_denominator <- local({
-    nd <- pop_df |>
-      dplyr::group_by(dplyr::across(dplyr::all_of(!!group_var))) |>
+    nd <- no_na_pop_df |>
+      dplyr::group_by(dplyr::across(dplyr::all_of(group_var))) |>
       dplyr::summarise(N = length(unique(.data[[subjid_var]])))
-    total <- length(unique(pop_df[[subjid_var]]))
+    total <- length(unique(no_na_pop_df[[subjid_var]]))
     stats::setNames(c(nd[["N"]], total), c(as.character(nd[[group_var]]), total_column_name))
   })
 
   reduced_group_event_df <- local({
     reduced_group_event_df <- event_df[, c(subjid_var, hierarchy), drop = FALSE]
     reduced_group_event_df <- dplyr::left_join(
-      reduced_group_event_df, pop_df[, c(subjid_var, group_var), drop = FALSE],
+      reduced_group_event_df, no_na_pop_df[, c(subjid_var, group_var), drop = FALSE],
       by = subjid_var
     )
   })
   # event_df may not contain all groups
-  levels(reduced_group_event_df[[group_var]]) <- levels(pop_df[[group_var]])
+  levels(reduced_group_event_df[[group_var]]) <- levels(no_na_pop_df[[group_var]])
 
   rev_hierarchy <- rev(hierarchy)
   hierarchy_df_list <- vector(mode = "list", length = length(rev_hierarchy))
@@ -362,7 +375,7 @@ sort_wide_format_event_table_to_HTML <- function(d, on_cell_click = NULL) { # no
       df[c(hierarchy, hier_lvl_col)], function(...) {
         args <- list(...)
         if (args[[hier_lvl_col]] > hierarchy_length) {
-          return("Overall")
+          return("Subjects with any event")
         }
         curr_lvl <- rev(hierarchy)[args[[hier_lvl_col]]]
         curr_label <- as.character(args[[curr_lvl]])
@@ -432,7 +445,7 @@ hierarchical_count_table_ui <- function(id) {
     shiny::div(
       shiny::div(style = "display: inline-block;", col_menu_UI(id = ns(EC$ID$HIERARCHY))),
       shiny::div(style = "display: inline-block;", col_menu_UI(id = ns(EC$ID$GRP))),
-      shiny::div(style = "display: inline-block;", shiny::numericInput(ns(EC$ID$MIN_PERCENT), label = "Minimum %", value = 0, min = 0, max = 100)),
+      shiny::div(style = "display: inline-block;", shiny::numericInput(ns(EC$ID$MIN_PERCENT), label = EC$LBL$MIN_PERCENT, value = 0, min = 0, max = 100)),
       shiny::div(style = "display: inline-block;", mod_export_counttable_UI(ns(EC$ID$TAB_DOWNLOAD)))
     ),
     shiny::uiOutput(ns(EC$ID$TABLE))
@@ -453,9 +466,9 @@ hierarchical_count_table_ui <- function(id) {
 #'
 #' @param subjid_var `character(1)`
 #' A string representing the subject identifier column in both datasets.
-#' 
+#'
 #' @param on_sbj_click_fun 'function()'
-#' 
+#'
 #' Function to invoke when a subject is clicked
 #'
 #' @param show_modal_on_click `logical(1)`
@@ -466,6 +479,12 @@ hierarchical_count_table_ui <- function(id) {
 #'
 #' @param default_group `character(1)|NULL`
 #' A default value for the group variable (optional).
+#'
+#' @param hierarchy_choices `character(1+)|NULL`
+#' A character vector specifying the possible choices for the hierarchy variables (optional).
+#'
+#' @param group_choices `character(1+)|NULL`
+#' A character vector specifying the possible choices for the group variable (optional).
 #'
 #' @param intended_use_label Either a string indicating the intended use for export, or
 #' NULL. The provided label will be displayed prior to the download and will also be included in the exported file.
@@ -485,6 +504,8 @@ hierarchical_count_table_server <- function(
     on_sbj_click_fun = function() NULL,
     default_hierarchy = NULL,
     default_group = NULL,
+    hierarchy_choices = NULL,
+    group_choices = NULL,
     intended_use_label = NULL) {
   mod <- function(input, output, session) {
     ns <- session[["ns"]]
@@ -492,21 +513,30 @@ hierarchical_count_table_server <- function(
     inputs <- list()
     inputs[[EC$ID$HIERARCHY]] <- col_menu_server(
       id = EC$ID$HIERARCHY, data = table_dataset,
-      label = "Event count by",
-      include_func = function(x) {
-        is.factor(x) || is.character(x)
+      label = shiny::div(shiny::tags$label(EC$LBL$HIERARCHY),
+                         shiny::icon("circle-info",
+                                     title = EC$INFO$HIERARCHY)),
+      include_func = function(var, var_name) {
+        (is.factor(var) || is.character(var)) &&
+          var_name != subjid_var &&
+          (is.null(hierarchy_choices) || var_name %in% hierarchy_choices)
       },
       default = default_hierarchy,
-      multiple = TRUE
+      multiple = TRUE,
+      include_none = FALSE,
+      options = list(maxItems = 2)
     )
 
     inputs[[EC$ID$GRP]] <- col_menu_server(
       id = EC$ID$GRP, data = pop_dataset,
-      label = "Group by",
-      include_func = function(x) {
-        is.factor(x) || is.character(x)
+      label = EC$LBL$GRP,
+      include_func = function(var, var_name) {
+        (is.factor(var) || is.character(var)) &&
+          var_name != subjid_var &&
+          (is.null(group_choices) || var_name %in% group_choices)
       },
-      default = default_group
+      default = default_group,
+      include_none = FALSE
     )
 
     inputs[[EC$ID$MIN_PERCENT]] <- shiny::reactive({
@@ -540,6 +570,10 @@ hierarchical_count_table_server <- function(
         shiny::need(
           checkmate::test_number(min_percent, na.ok = FALSE, lower = 0, upper = 100),
           EC$MSG$VALIDATE$NO_MIN_PERCENT
+        ),
+        shiny::need(
+          !checkmate::test_choice(group_var, hierarchy, null.ok = TRUE),
+          EC$MSG$VALIDATE$GRP_CLASH
         )
       )
 
@@ -596,10 +630,10 @@ hierarchical_count_table_server <- function(
       })
     }
 
-    # Jumping and communication    
+    # Jumping and communication
     shiny::observeEvent(input[["clicked_sbj"]], {
       shiny::req(checkmate::test_string(input[["clicked_sbj"]], na.ok = FALSE, min.chars = 1, null.ok = FALSE))
-      shiny::removeModal()     
+      shiny::removeModal()
       on_sbj_click_fun()
     })
 
@@ -636,9 +670,9 @@ hierarchical_count_table_server <- function(
 #' Dataset dispatcher. This parameter is incompatible with its *_dataset_name counterpart. Only for advanced use.
 #'
 #' @param receiver_id `character(1)`
-#' 
-#' Shiny ID of the module receiving the selected subject ID in the data listing. This ID must be present in the app or be NULL. 
-#' 
+#'
+#' Shiny ID of the module receiving the selected subject ID in the data listing. This ID must be present in the app or be NULL.
+#'
 #'
 #' @keywords main
 #'
@@ -650,6 +684,8 @@ mod_hierarchical_count_table <- function(module_id,
                                          show_modal_on_click = TRUE,
                                          default_hierarchy = NULL,
                                          default_group = NULL,
+                                         hierarchy_choices = NULL,
+                                         group_choices = NULL,
                                          intended_use_label = "Use only for internal review and monitoring during the conduct of clinical trials.",
                                          receiver_id = NULL) {
   mod <- list(
@@ -672,6 +708,8 @@ mod_hierarchical_count_table <- function(module_id,
         on_sbj_click_fun = on_sbj_click_fun,
         default_hierarchy = default_hierarchy,
         default_group = default_group,
+        hierarchy_choices = hierarchy_choices,
+        group_choices = group_choices,
         intended_use_label = intended_use_label
       )
     },
@@ -691,6 +729,8 @@ mod_hierarchical_count_table_API_docs <- list(
   show_modal_on_click = "",
   default_hierarchy = "",
   default_group = "",
+  hierarchy_choices = "",
+  group_choices = "",
   intended_use_label = "",
   receiver_id = ""
 )
@@ -704,6 +744,10 @@ mod_hierarchical_count_table_API_spec <- TC$group(
   default_hierarchy = TC$col("table_dataset_name", TC$or(TC$character(), TC$factor())) |>
     TC$flag("zero_or_more", "optional"),
   default_group = TC$col("pop_dataset_name", TC$or(TC$character(), TC$factor())) |> TC$flag("optional"),
+  hierarchy_choices = TC$col("table_dataset_name", TC$or(TC$character(), TC$factor())) |>
+    TC$flag("zero_or_more", "optional"),
+  group_choices = TC$col("pop_dataset_name", TC$or(TC$character(), TC$factor())) |>
+    TC$flag("zero_or_more", "optional"),
   intended_use_label = TC$character() |> TC$flag("optional"),
   receiver_id = TC$character() |> TC$flag("optional")
 ) |> TC$attach_docs(mod_hierarchical_count_table_API_docs)
@@ -711,7 +755,9 @@ mod_hierarchical_count_table_API_spec <- TC$group(
 
 check_mod_hierarchical_count_table <- function(
     afmm, datasets, module_id, table_dataset_name, pop_dataset_name, subjid_var, show_modal_on_click,
-    default_hierarchy, default_group, intended_use_label, receiver_id) {
+    default_hierarchy, default_group,
+    hierarchy_choices, group_choices,
+    intended_use_label, receiver_id) {
   warn <- CM$container()
   err <- CM$container()
 
@@ -721,7 +767,9 @@ check_mod_hierarchical_count_table <- function(
   OK <- check_mod_hierarchical_count_table_auto( # nolint unused
     afmm, datasets,
     module_id, table_dataset_name, pop_dataset_name, subjid_var, show_modal_on_click,
-    default_hierarchy, default_group, intended_use_label, receiver_id,
+    default_hierarchy, default_group,
+    hierarchy_choices, group_choices,
+    intended_use_label, receiver_id,
     warn, err
   )
 
@@ -828,7 +876,7 @@ mock_app_hierarchical_count_table_mm <- function() { # nolint
 
   dv.manager::run_app(
     data = list(
-      dummy = list(adae = pharmaverseadam::adae |> chr2factor(), adsl = pharmaverseadam::adsl |> chr2factor())
+      dummy = list(adae = pharmaverseadam::adae, adsl = pharmaverseadam::adsl)
     ),
     module_list = list(
       "ADAE by term" = mod_hierarchical_count_table(