+++ R/utils-misc.R
@@ -34,4 +34,23 @@ ssub <- function(s, ...) {
   s
 }
 
+#' Replace NA values in a factor with character value and add associated level
+#'
+#' @param f Factor
+#' @param level New level to associate with original NA values
+#'
+#' @return A factor with NA replaced with character value/level
+#'
+#' @keywords internal
+add_na_factor_level <- function(f, level = "<NA>") {
+
+  if (!any(is.na(f))) return(f)
+
+  ch <- tidyr::replace_na(as.character(f), level)
+  new_levels <- c(levels(f), level)
+  new_factor <- factor(ch, levels = new_levels)
+
+  return(new_factor)
+}
+
 if_not_null <- Negate(is.null)