+++ NEWS.md
@@ -1,4 +1,12 @@
+# dv.tables 0.0.5-9000
+
+* Allow subsetting of hierarchy and group variables drop-down list through module argument
+* Remove subject identifier variable from hierarchy and group variables drop-down list
+* Alert when group selection also selected in hierarchy
+* Allow a maximum of 2 hierarchy variable selections
+
 # dv.tables 0.0.4-9000
+
 * Fix extra line due to truncation
 
 # dv.tables 0.0.3.9000